bar plot enhanced for putting color on bars according to the belonging state of the port
Added some comments. 
Added a way to use the lib colorcet (https://colorcet.holoviz.org/getting_started/index.html#usage) in bokeh, allowing for a dynamic choice of the number of the colors (= the number of different states) 
Legend also is modified to be displayed with all the items
Ports are sorted according their state and their name.